api/products/requirements.txt
fastapi==0.115.0
httpx==0.27.2
motor==3.6.0
pydantic-settings==2.5.2
PyJWT==2.9.0
pyjwt[crypto]